Fix bug that made text color and/or style spread in some cases
Fixes #2

When `QTextBrowser.clear` is called, it normally resets the current char
format as well. However, the format was not being cleared because of the
`tb.scrollToAnchor("top")` in `set_QTextBrowser_text` which put the
cursor at the beginning of the document (clear is designed to not reset
the format in this situation). When text was selected in the browser,
the current char format becomes the format of the end of the selection.